[sonataflow-operator] Knative outdated revisions are not cleaned-up for workflows with the knative deploymentModel and gitops profile
