+++ packages/sonataflow-operator/internal/controller/profiles/gitops/states_gitops.go
@@ -19,6 +19,9 @@ package gitops
 
 import (
 	"context"
+	"fmt"
+
+	"github.com/apache/incubator-kie-tools/packages/sonataflow-operator/internal/controller/knative"
 
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -69,6 +72,9 @@ func (f *followDeployWorkflowState) Do(ctx context.Context, workflow *operatorap
 }
 
 func (f *followDeployWorkflowState) PostReconcile(ctx context.Context, workflow *operatorapi.SonataFlow) error {
-	//By default, we don't want to perform anything after the reconciliation, and so we will simply return no error
+	// Clean up the outdated Knative revisions, if any
+	if err := knative.CleanupOutdatedRevisions(ctx, f.C, f.Cfg, workflow); err != nil {
+		return fmt.Errorf("failied to cleanup workflow outdated revisions, workflow: %s, namespace: %s - %v", workflow.Name, workflow.Namespace, err)
+	}
 	return nil
 }
